ISAICP-2266-BACKPORT: Finalize restructure of tests. There will now be a Node test base for community content.
